+++ curve25519-dalek/src/edwards.rs
@@ -103,7 +103,10 @@ use core::ops::{AddAssign, SubAssign};
 use core::ops::{Mul, MulAssign};
 
 #[cfg(feature = "digest")]
-use digest::{generic_array::typenum::U64, Digest};
+use digest::{
+    consts::True, crypto_common::BlockSizeUser, generic_array::typenum::U64, typenum::IsGreater,
+    Digest, FixedOutput, HashMarker,
+};
 
 #[cfg(feature = "group")]
 use {
@@ -257,6 +260,8 @@ impl TryFrom<&[u8]> for CompressedEdwardsY {
 // structs containing `EdwardsPoint`s and use Serde's derived
 // serializers to serialize those structures.
 
+#[cfg(feature = "digest")]
+use constants::ED25519_SQRTAM2;
 #[cfg(feature = "serde")]
 use serde::de::Visitor;
 #[cfg(feature = "serde")]
@@ -620,11 +625,61 @@ impl EdwardsPoint {
     }
 
     #[cfg(feature = "digest")]
+    /// Perform hashing to curve, with explicit hash function and domain separator, `domain_sep`,
+    /// using the suite `edwards25519_XMD:SHA-512_ELL2_NU_`. The input is the concatenation of the
+    /// elements of `bytes`. Likewise for the domain separator with `domain_sep`. At least one
+    /// element of `domain_sep`, MUST be nonempty, and the concatenation MUST NOT exceed
+    /// 255 bytes.
+    ///
+    /// # Panics
+    /// Panics if `domain_sep.collect().len() == 0` or `> 255`
+    pub fn hash_to_curve<D>(bytes: &[&[u8]], domain_sep: &[&[u8]]) -> EdwardsPoint
+    where
+        D: BlockSizeUser + Default + FixedOutput<OutputSize = U64> + HashMarker,
+        D::BlockSize: IsGreater<D::OutputSize, Output = True>,
+    {
+        // For reference see
+        // https://www.rfc-editor.org/rfc/rfc9380.html#name-elligator-2-method-2
+
+        let fe = FieldElement::hash_to_field::<D>(bytes, domain_sep);
+        let (M1, is_sq) = crate::montgomery::elligator_encode(&fe);
+
+        // The `to_edwards` conversion we're performing takes as input the sign of the Edwards
+        // `y` coordinate. However, the specification uses `is_sq` to determine the sign of the
+        // Montgomery `v` coordinate. Our approach reconciles this mismatch as follows:
+        //
+        // * We arbitrarily fix the sign of the Edwards `y` coordinate (we choose 0).
+        // * Using the Montgomery `u` coordinate and the Edwards `X` coordinate, we recover `v`.
+        // * We verify that the sign of `v` matches the expected one, i.e., `is_sq == mont_v.is_negative()`.
+        // * If it does not match, we conditionally negate to correct the sign.
+        //
+        // Note: This logic aligns with the RFC draft specification:
+        //     https://www.rfc-editor.org/rfc/rfc9380.html#name-elligator-2-method-2
+        // followed by the mapping
+        //     https://www.rfc-editor.org/rfc/rfc9380.html#name-mappings-for-twisted-edward
+        // The only difference is that our `elligator_encode` returns only the Montgomery `u` coordinate,
+        // so we apply this workaround to reconstruct and validate the sign.
+
+        let mut E1_opt = M1
+            .to_edwards(0)
+            .expect("Montgomery conversion to Edwards point in Elligator failed");
+
+        // Now we recover v, to ensure that we got the sign right.
+        let mont_v =
+            &(&ED25519_SQRTAM2 * &FieldElement::from_bytes(&M1.to_bytes())) * &E1_opt.X.invert();
+        E1_opt.X.conditional_negate(is_sq ^ mont_v.is_negative());
+        E1_opt.mul_by_cofactor()
+    }
+
+    #[cfg(feature = "digest")]
     /// Maps the digest of the input bytes to the curve. This is NOT a hash-to-curve function, as
     /// it produces points with a non-uniform distribution. Rather, it performs something that
     /// resembles (but is not) half of the
-    /// [`hash_to_curve`](https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-16.html#section-3-4.2.1)
+    /// [`hash_to_curve`](https://www.rfc-editor.org/rfc/rfc9380.html#section-3-4.2.1)
     /// function from the Elligator2 spec.
+    ///
+    /// For a hash to curve with uniform distribution and compatible with the spec, see
+    /// [`Self::hash_to_curve`].
     #[deprecated(
         since = "4.0.0",
         note = "previously named `hash_from_bytes`, this is not a secure hash function"
@@ -643,7 +698,7 @@ impl EdwardsPoint {
 
         let fe = FieldElement::from_bytes(&res);
 
-        let M1 = crate::montgomery::elligator_encode(&fe);
+        let (M1, _) = crate::montgomery::elligator_encode(&fe);
         let E1_opt = M1.to_edwards(sign_bit);
 
         E1_opt
@@ -2358,7 +2413,7 @@ mod test {
     ////////////////////////////////////////////////////////////
 
     #[cfg(all(feature = "alloc", feature = "digest"))]
-    fn test_vectors() -> Vec<Vec<&'static str>> {
+    fn signal_test_vectors() -> Vec<Vec<&'static str>> {
         vec![
             vec![
                 "214f306e1576f5a7577636fe303ca2c625b533319f52442b22a9fa3b7ede809f",
@@ -2407,12 +2462,79 @@ mod test {
     #[allow(deprecated)]
     #[cfg(all(feature = "alloc", feature = "digest"))]
     fn elligator_signal_test_vectors() {
-        for vector in test_vectors().iter() {
+        for vector in signal_test_vectors().iter() {
             let input = hex::decode(vector[0]).unwrap();
             let output = hex::decode(vector[1]).unwrap();
 
             let point = EdwardsPoint::nonspec_map_to_curve::<sha2::Sha512>(&input);
             assert_eq!(point.compress().to_bytes(), output[..]);
+        }
+    }
+
+    // Hash-to-curve test vectors from
+    // https://www.rfc-editor.org/rfc/rfc9380.html#name-edwards25519_xmdsha-512_ell2
+    // These are of the form (input_msg, output_x, output_y)
+    #[cfg(all(feature = "alloc", feature = "digest"))]
+    const RFC_HASH_TO_CURVE_KAT: &[(&[u8], &str, &str)] = &[
+        (
+            b"",
+            "1ff2b70ecf862799e11b7ae744e3489aa058ce805dd323a936375a84695e76da",
+            "222e314d04a4d5725e9f2aff9fb2a6b69ef375a1214eb19021ceab2d687f0f9b",
+        ),
+        (
+            b"abc",
+            "5f13cc69c891d86927eb37bd4afc6672360007c63f68a33ab423a3aa040fd2a8",
+            "67732d50f9a26f73111dd1ed5dba225614e538599db58ba30aaea1f5c827fa42",
+        ),
+        (
+            b"abcdef0123456789",
+            "1dd2fefce934ecfd7aae6ec998de088d7dd03316aa1847198aecf699ba6613f1",
+            "2f8a6c24dd1adde73909cada6a4a137577b0f179d336685c4a955a0a8e1a86fb",
+        ),
+        (
+            b"q128_qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq\
+            qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq",
+            "35fbdc5143e8a97afd3096f2b843e07df72e15bfca2eaf6879bf97c5d3362f73",
+            "2af6ff6ef5ebba128b0774f4296cb4c2279a074658b083b8dcca91f57a603450",
+        ),
+        (
+            b"a512_aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\
+            aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\
+            aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\
+            aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\
+            aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\
+            aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+            "6e5e1f37e99345887fc12111575fc1c3e36df4b289b8759d23af14d774b66bff",
+            "2c90c3d39eb18ff291d33441b35f3262cdd307162cc97c31bfcc7a4245891a37"
+        )
+    ];
+
+    #[test]
+    #[cfg(all(feature = "alloc", feature = "digest"))]
+    fn elligator_hash_to_curve_test_vectors() {
+        let dst = b"QUUX-V01-CS02-with-edwards25519_XMD:SHA-512_ELL2_NU_";
+        for (index, vector) in RFC_HASH_TO_CURVE_KAT.iter().enumerate() {
+            let input = vector.0;
+
+            let expected_output = {
+                let mut x_bytes = hex::decode(vector.1).unwrap();
+                x_bytes.reverse();
+                let x = FieldElement::from_bytes(&x_bytes.try_into().unwrap());
+
+                let mut y_bytes = hex::decode(vector.2).unwrap();
+                y_bytes.reverse();
+                let y = FieldElement::from_bytes(&y_bytes.try_into().unwrap());
+
+                EdwardsPoint {
+                    X: x,
+                    Y: y,
+                    Z: FieldElement::ONE,
+                    T: &x * &y,
+                }
+            };
+
+            let computed = EdwardsPoint::hash_to_curve::<sha2::Sha512>(&[&input], &[dst]);
+            assert_eq!(computed, expected_output, "Failed in test {}", index);
         }
     }
 }